requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.8
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-choices==1.7.1
djangorestframework==3.11.0
drf-yasg==1.17.1
flower==0.9.5
future==0.18.2
humanize==2.5.0
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.10
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
prometheus-client==0.8.0
psycopg2==2.8.5
pyparsing==2.4.7
pyTelegramBotAPI==3.7.2
python-memcached==1.59
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.19.1
vine==1.3.0
